Set the bright foreground color in gnome-dark
My original intention was that the gnome-dark scheme would use the colour corresponding to 'bright white' as the bright version of the foreground colour. (The default foreground colour is 'normal white'.) This didn't make it into the gnome-terminal code, because the colour scheme settings are not able to set the bright foreground colour. But it can be set here.